requirement.txt
numpy>=1.22.0
scipy>=1.8.0
scikit-learn>=1.0.2
matplotlib>=3.5.0
Theano-PyMC>=1.1.2
Lasagne==0.2.dev1